--- app/src/main/java/com/android/counter/App.kt
@@ -1,22 +1,0 @@
-package com.android.counter
-
-import android.app.Application
-import com.github.terrakok.cicerone.Cicerone
-import com.github.terrakok.cicerone.Router
-
-class App : Application() {
-    companion object {
-        lateinit var instance: App
-    }
-
-    private val cicerone: Cicerone<Router> by lazy {
-        Cicerone.create()
-    }
-    val navigatorHolder get() = cicerone.getNavigatorHolder()
-    val router get() = cicerone.router
-
-    override fun onCreate() {
-        super.onCreate()
-        instance = this
-    }
-}
