--- app/src/main/java/com/android/counter/mvp/model/GithubUsersRepo.kt
@@ -1,17 +1,0 @@
-package com.android.counter.mvp.model
-
-import com.android.counter.mvp.model.entity.GithubUser
-
-class GithubUsersRepo {
-    private val repositories = listOf(
-        GithubUser("login1"),
-        GithubUser("login2"),
-        GithubUser("login3"),
-        GithubUser("login4"),
-        GithubUser("login5")
-    )
-
-    fun getUsers(): List<GithubUser> {
-        return repositories
-    }
-}
