--- app/src/main/java/com/android/counter/mvp/model/entity/GithubUser.kt
@@ -1,9 +1,0 @@
-package com.android.counter.mvp.model.entity
-
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
-
-@Parcelize
-data class GithubUser(
-    val login: String,
-) : Parcelable
