--- app/src/main/java/com/android/counter/mvp/presenter/MainPresenter.kt
@@ -1,19 +1,0 @@
-package com.android.counter.mvp.presenter
-
-import com.android.counter.mvp.view.MainView
-import com.android.counter.navigation.IScreens
-import com.github.terrakok.cicerone.Router
-import moxy.MvpPresenter
-
-class MainPresenter(val router: Router, val screens: IScreens) :
-    MvpPresenter<MainView>() {
-
-    override fun onFirstViewAttach() {
-        super.onFirstViewAttach()
-        router.replaceScreen(screens.users())
-    }
-
-    fun backClicked() {
-        router.exit()
-    }
-}
