--- app/src/main/java/com/android/counter/mvp/presenter/list/IListPresenter.kt
@@ -1,9 +1,0 @@
-package com.android.counter.mvp.presenter.list
-
-import com.android.counter.mvp.view.list.IItemView
-
-interface IListPresenter<V : IItemView> {
-    var itemClickListener: ((V) -> Unit)?
-    fun bindView(view: V)
-    fun getCount(): Int
-}
