--- app/src/main/java/com/android/counter/mvp/presenter/list/IUserListPresenter.kt
@@ -1,5 +1,0 @@
-package com.android.counter.mvp.presenter.list
-
-import com.android.counter.mvp.view.list.IUserItemView
-
-interface IUserListPresenter : IListPresenter<IUserItemView>
