--- app/src/main/java/com/android/counter/mvp/view/MainView.kt
@@ -1,8 +1,0 @@
-package com.android.counter.mvp.view
-
-import moxy.MvpView
-import moxy.viewstate.strategy.AddToEndSingleStrategy
-import moxy.viewstate.strategy.StateStrategyType
-
-@StateStrategyType(AddToEndSingleStrategy::class)
-interface MainView : MvpView
