--- app/src/main/java/com/android/counter/mvp/view/UsersView.kt
@@ -1,13 +1,0 @@
-package com.android.counter.mvp.view
-
-import moxy.MvpView
-import moxy.viewstate.strategy.AddToEndSingleStrategy
-import moxy.viewstate.strategy.StateStrategyType
-
-@StateStrategyType(AddToEndSingleStrategy::class)
-interface UsersView : MvpView {
-    fun init()
-    fun updateList()
-    fun release()
-}
-
