--- app/src/main/java/com/android/counter/mvp/view/list/IItemView.kt
@@ -1,6 +1,0 @@
-package com.android.counter.mvp.view.list
-
-interface IItemView {
-    var pos: Int
-}
-
