--- app/src/main/java/com/android/counter/mvp/view/list/IUserItemView.kt
@@ -1,5 +1,0 @@
-package com.android.counter.mvp.view.list
-
-interface IUserItemView  : IItemView {
-    fun setLogin(text: String)
-}
