--- app/src/main/java/com/android/counter/navigation/AndroidScreens.kt
@@ -1,8 +1,0 @@
-package com.android.counter.navigation
-
-import com.android.counter.ui.fragment.UsersFragment
-import com.github.terrakok.cicerone.androidx.FragmentScreen
-
-class AndroidScreens : IScreens {
-    override fun users() = FragmentScreen { UsersFragment.newInstance() }
-}
