--- app/src/main/java/com/android/counter/navigation/IScreens.kt
@@ -1,7 +1,0 @@
-package com.android.counter.navigation
-
-import com.github.terrakok.cicerone.Screen
-
-interface IScreens {
-    fun users(): Screen
-}
