--- app/src/main/java/com/android/counter/ui/activity/BackButtonListener.kt
@@ -1,5 +1,0 @@
-package com.android.counter.ui.activity
-
-interface BackButtonListener {
-    fun backPressed(): Boolean
-}
