--- app/src/main/java/com/android/counter/ui/activity/MainActivity.kt
@@ -1,48 +1,0 @@
-package com.android.counter.ui.activity
-
-import android.os.Bundle
-import com.android.counter.R
-import com.android.counter.databinding.ActivityMainBinding
-import com.android.counter.mvp.presenter.MainPresenter
-import com.android.counter.mvp.view.MainView
-import com.android.counter.navigation.AndroidScreens
-import com.android.counter.App
-import com.github.terrakok.cicerone.androidx.AppNavigator
-import moxy.MvpAppCompatActivity
-import moxy.ktx.moxyPresenter
-
-class MainActivity : MvpAppCompatActivity(), MainView {
-    private lateinit var binding: ActivityMainBinding
-
-    val navigator = AppNavigator(this, R.id.container)
-
-    private val presenter by moxyPresenter {
-        MainPresenter(App.instance.router, AndroidScreens())
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-
-        setContentView(binding.root)
-    }
-
-    override fun onResumeFragments() {
-        super.onResumeFragments()
-        App.instance.navigatorHolder.setNavigator(navigator)
-    }
-
-    override fun onPause() {
-        super.onPause()
-        App.instance.navigatorHolder.removeNavigator()
-    }
-
-    override fun onBackPressed() {
-        supportFragmentManager.fragments.forEach {
-            if (it is BackButtonListener && it.backPressed()) {
-                return
-            }
-        }
-        presenter.backClicked()
-    }
-}
