+++ app/src/main/java/com/android/githubclient/mvp/model/GithubUsersRepo.kt
@@ -1,0 +1,17 @@
+package com.android.githubclient.mvp.model
+
+import com.android.githubclient.mvp.model.entity.GithubUser
+
+class GithubUsersRepo {
+    private val repositories = listOf(
+        GithubUser("login1"),
+        GithubUser("login2"),
+        GithubUser("login3"),
+        GithubUser("login4"),
+        GithubUser("login5")
+    )
+
+    fun getUsers(): List<GithubUser> {
+        return repositories
+    }
+}
