+++ app/src/main/java/com/android/githubclient/mvp/model/entity/GithubUser.kt
@@ -1,0 +1,10 @@
+package com.android.githubclient.mvp.model.entity
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Suppress("DEPRECATED_ANNOTATION")
+@Parcelize
+data class GithubUser(
+    val login: String,
+) : Parcelable
