+++ app/src/main/java/com/android/githubclient/mvp/presenter/MainPresenter.kt
@@ -1,0 +1,19 @@
+package com.android.githubclient.mvp.presenter
+
+import com.android.githubclient.mvp.view.MainView
+import com.android.githubclient.navigation.IScreens
+import com.github.terrakok.cicerone.Router
+import moxy.MvpPresenter
+
+class MainPresenter(val router: Router, val screens: IScreens) :
+    MvpPresenter<MainView>() {
+
+    override fun onFirstViewAttach() {
+        super.onFirstViewAttach()
+        router.replaceScreen(screens.users())
+    }
+
+    fun backClicked() {
+        router.exit()
+    }
+}
