+++ app/src/main/java/com/android/githubclient/mvp/presenter/list/IUserListPresenter.kt
@@ -1,0 +1,5 @@
+package com.android.githubclient.mvp.presenter.list
+
+import com.android.githubclient.mvp.view.list.IUserItemView
+
+interface IUserListPresenter : IListPresenter<IUserItemView>
