+++ app/src/main/java/com/android/githubclient/mvp/view/MainView.kt
@@ -1,0 +1,8 @@
+package com.android.githubclient.mvp.view
+
+import moxy.MvpView
+import moxy.viewstate.strategy.AddToEndSingleStrategy
+import moxy.viewstate.strategy.StateStrategyType
+
+@StateStrategyType(AddToEndSingleStrategy::class)
+interface MainView : MvpView
