+++ app/src/main/java/com/android/githubclient/mvp/view/UsersView.kt
@@ -1,0 +1,13 @@
+package com.android.githubclient.mvp.view
+
+import moxy.MvpView
+import moxy.viewstate.strategy.AddToEndSingleStrategy
+import moxy.viewstate.strategy.StateStrategyType
+
+@StateStrategyType(AddToEndSingleStrategy::class)
+interface UsersView : MvpView {
+    fun init()
+    fun updateList()
+    fun release()
+}
+
