+++ app/src/main/java/com/android/githubclient/mvp/view/list/IItemView.kt
@@ -1,0 +1,6 @@
+package com.android.githubclient.mvp.view.list
+
+interface IItemView {
+    var pos: Int
+}
+
