+++ app/src/main/java/com/android/githubclient/mvp/view/list/IUserItemView.kt
@@ -1,0 +1,5 @@
+package com.android.githubclient.mvp.view.list
+
+interface IUserItemView  : IItemView {
+    fun setLogin(text: String)
+}
