+++ app/src/main/java/com/android/githubclient/ui/activity/BackButtonListener.kt
@@ -1,0 +1,5 @@
+package com.android.githubclient.ui.activity
+
+interface BackButtonListener {
+    fun backPressed(): Boolean
+}
