+++ app/src/main/res/layout/fragment_users.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_users"
@@ -10,4 +11,5 @@
         android:clipToPadding="false"
         android:paddingTop="16dp"
         android:paddingBottom="16dp" />
-</RelativeLayout>
+
+</LinearLayout>
