+++ settings.gradle
@@ -13,5 +13,5 @@ dependencyResolutionManagement {
         jcenter()
     }
 }
-rootProject.name = "Counter"
+rootProject.name = "GithubClient"
 include ':app'
